Rename Mac -> Macro
This is part of a move to standardize on abbreviating only those types in the
AST that are abbreviated in Rust source code, like `impl` or `fn`. The grammar
of Rust source code never abbreviates `macro` to `mac`, so this commit reflects
that.

Field names are being kept as `mac` because `macro` is a reserved word.