Flatten ForeignItem into an enum
There is some duplication here but it is worth it for having foreign items
individually parsable and printable with Synom and ToTokens impls. This is also
more flexible as we add new types of foreign item, for example macros that can
expand to foreign items. A macro foreign item would not have a visibility and
some of the other fields previously held in the ForeignItem struct.