Add an `Error::new_spanned` constructor
This commit is intended to expose the ability to have multi-token errors
on stable Rust right now. This commit also, though, hopes that `syn` can
largely stick to "one `Span` is all you need" everywhere for the time
being. The implementation here uses two spans internally for a
start/end, but that's not exposed through the API at all yet.

Closes #537