Make turbofish optional
Implements RFC 2544 which is in proposed-FCP.

This simple implementation works but diagnostics are somewhat worse than
before. Consider the following input in which the user is missing part
of the second type parameter. The parser backtracks to reparse the first
type parameter as an expression, concluding that it is supposed to be
something like `&('a: loop {})` and suggesting that the user write the
missing colon after their loop label.

We may be able to build out a system for collecting both errors and
emitting whichever one happened after parsing the larger number of
tokens.

Before:

    error: expected identifier
     --> src/main.rs:4:38
      |
    4 |     let _ = f::<&'a fn(), for<'a> serde::>();
      |                                          ^

After, without turbofish:

    error: expected `:`
     --> src/main.rs:4:19
      |
    4 |     let _ = f<&'a fn(), for<'a> serde::>();
      |                   ^^