Rewrite codegen
The primary purpose is to split up parsing the `syn` source from
generating visit / visit_mut / fold.

The commit defines an intermediate representation of syn's internal
types. The first step is then to parse the syn source and construct the
intermediate representation. Then, code is generated using the IR.

This change is a precursor to generating a JSON version of the IR.