Prodvide a cfg for the doc-cfg feature
To be able to include banners, which feature is needed for an item, the
feature `#![feature(doc_cfg)]` has to be available. This feture is only
available in nightly compiler since August 2017. So we need to detect,
whether we are building the docs under a nightly compiler to add the
doc-cfg-banners.