Make the group::parse_group public
I am working on building a substitute `Expr` enum which only allows a subset of Rust expressions. I was able to finish off it except for this one small thing that I needed to use.

This small function contains a lot of code I don't want to duplication since it doesn't depend on `Expr`. If you make this `pub`, my use case would be satisfied.

Can you please merge and release a new version ASAP? Thanks.