Allow parsing `MetaNameValue` if the path part contains a keyword
This allows this to be parsed:

```rust
let attr: syn::MetaNameValue = syn::parse_quote!{
    as = "FooBar"
};
```

This aligns the code with `parse_meta_path` which already uses
`input.peek(Ident::peek_any)`.
https://github.com/dtolnay/syn/blob/25def51081b68538547f34a3494744a807be4203/src/meta.rs#L401

This allows parsing the `#[serde_as(as = "FooBar")]` attributes which
are used as part of the `serde_with` crate.

Thanks to @matthias-stemmler for finding the cause.
https://github.com/dtolnay/syn/issues/1408#issuecomment-1476691276